requirements.txt
# SIAMS Web Dashboard Dependencies
# For Streamlit deployment (Azure/Cloud)

streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
pytz>=2023.3
requests>=2.31.0

# AI/LLM dependencies (with fallback)
google-generativeai>=0.8.5

# For data visualization (if needed)
plotly>=5.17.0
matplotlib>=3.7.0
